+++ sdk/core/azure-core/inc/azure/core/credentials.hpp
@@ -16,7 +16,6 @@
 #include <mutex>
 #include <string>
 #include <utility>
-#include <vector>
 
 namespace Azure { namespace Core {
 
@@ -36,6 +35,10 @@ namespace Azure { namespace Core {
     DateTime ExpiresOn;
   };
 
+  namespace Http {
+    struct TokenRequestOptions;
+  } // namespace Http
+
   /**
    * @brief Token credential.
    */
@@ -45,10 +48,11 @@ namespace Azure { namespace Core {
      * @brief Get an authentication token.
      *
      * @param context #Context so that operation can be cancelled.
-     * @param scopes Authentication scopes.
+     * @param tokenRequestOptions Options to get the token.
      */
-    virtual AccessToken GetToken(Context const& context, std::vector<std::string> const& scopes)
-        const = 0;
+    virtual AccessToken GetToken(
+        Context const& context,
+        Http::TokenRequestOptions const& tokenRequestOptions) const = 0;
 
     /// Destructor.
     virtual ~TokenCredential() = default;