+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0-beta.3 (Unreleased)
 
+### Breaking Changes
+
+- `ClientSecretCredential ` constructor takes `ClientSecretCredentialOptions` struct instead of authority host string. `TokenCredentialOptions` struct has authority host string as data member.
 
 ## 1.0.0-beta.2 (2021-01-13)